Add operator overloads for enMuitiViewType enums.

+++ Graphcis_Framework_Neon/Graphcis_Framework_Neon/src/enSwapChain.cpp
@@ -123,8 +123,9 @@ enSwapChain::ResizeSwapChain(enWindow& currentWindow,
 
   deviceContext.OMSetRenderTargets(&renderTargetView, &DepthStencilView);
 
-  viewport.m_descriptor.width = newSize.x;
-  viewport.m_descriptor.height= newSize.y;
+  viewport.m_descriptor.width = static_cast< uint32 >(newSize.x);
+
+  viewport.m_descriptor.height = static_cast< uint32 >(newSize.y);
   
   deviceContext.RSSetViewports(&viewport);
